Refactor StateMap
1. Remove mutex: We only need to access StateMap from the main thread.

2. Remove `PARENT_FAILURE` state: This value is only read, never
written. This is a bug, but we plan to reintroduce it at a later date
with a better approach.

3. Make `StateMap` subclass `std::vector`. It's still a distinct class
for now.

4. Make `realisation_state_enum` into a strict C++11 struct called `State`

5. Rename Python's `RealizationStateEnum` into `State`, remove the
prefix `STATE_`.

6. Remove `is_legal_transition`: Any valid enum value is now legal and
it's up to the caller to make the correct decision.

7. Replace `createMask`/`selectMatching` with `indices_with_data`: The
only real use-case was creating a list of indices which have the state
`State.HAS_DATA`. The generic `__iter__` is available for other
use-cases.

+++ src/clib/lib/enkf/enkf_obs.cpp
@@ -269,7 +269,7 @@ int enkf_obs_get_size(const enkf_obs_type *obs) {
 
 static void enkf_obs_get_obs_and_measure_summary(
     const enkf_obs_type *enkf_obs, obs_vector_type *obs_vector,
-    enkf_fs_type *fs, const std::vector<int> &ens_active_list,
+    enkf_fs_type *fs, const std::vector<size_t> &ens_active_list,
     meas_data_type *meas_data, obs_data_type *obs_data) {
 
     int active_count = 0;
@@ -356,7 +356,7 @@ static void enkf_obs_get_obs_and_measure_summary(
 
 static void enkf_obs_get_obs_and_measure_node(
     const enkf_obs_type *enkf_obs, enkf_fs_type *fs, std::string obs_key,
-    const std::vector<int> &ens_active_list, meas_data_type *meas_data,
+    const std::vector<size_t> &ens_active_list, meas_data_type *meas_data,
     obs_data_type *obs_data) {
 
     obs_vector_type *obs_vector =
@@ -395,7 +395,7 @@ static void enkf_obs_get_obs_and_measure_node(
 void enkf_obs_get_obs_and_measure_data(
     const enkf_obs_type *enkf_obs, enkf_fs_type *fs,
     const std::vector<std::pair<std::string, std::vector<int>>> &observations,
-    const std::vector<int> &ens_active_list, meas_data_type *meas_data,
+    const std::vector<size_t> &ens_active_list, meas_data_type *meas_data,
     obs_data_type *obs_data) {
 
     for (auto &observation : observations) {
